Added health and damage plugins

+++ src/game/player/resources.rs
@@ -2,7 +2,7 @@ use bevy::prelude::*;
 
 #[derive(Resource)]
 pub struct Health {
-    pub amount: u8,
+    pub amount: u32,
 }
 
 impl Default for Health {
